ecsalt (ecsalt v2.0.0)

Entity Component System

Types

component()

-type component() :: {term(), term()}.

entity()

-type entity() :: {term(), [component()]}.

world()

-opaque world()

Functions

delete/1

-spec delete(world()) -> ok.

Stop the ECS and delete the ETS tables

new()

-spec new() -> world().

Start the ECS with ETS tables set to public, return them in an opaque world() type to the caller

proc(Data, World)

-spec proc(term(), world()) -> [any()].

For every registered system, trigger the proc callback with some extra data specified by Data and gather results.

to_map/1

-spec to_map(entity()) -> map().

Convert the component list for a given entity ID to a map. Warning: this can potentially create an unbounded number of atoms!